Add files via upload
データベースの起動をonClickで起動するようにしました。カレンダーアクティビティからフラグメントにタップした日付の値が移るようにプログラムを追加途中です。

+++ main/java/com/example/kakeibo/database/Database.java
@@ -1,5 +1,6 @@
 package com.example.kakeibo.database;
 
+import android.app.Activity;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
@@ -18,18 +19,20 @@ public class Database {
         return database;
     }
 
+    //書き込み用のSQLite取得
     private Database(Context context){
         helper = new TestOpenHelper(context);
         db = helper.getWritableDatabase();
     }
 
-    public boolean addEntry(String category, long price){
+    public boolean addEntry(String category, int price){
         final ContentValues values = new ContentValues();
         values.put(TestOpenHelper.COLUMN_CATEGORY, category);
         values.put(TestOpenHelper.COLUMN_PRICE,price);
         return db.insert(TestOpenHelper.TABLE_NAME, null, values) != -1;
     }
 
+    //データベースの読み込み
     public Cursor retrieveAllEntries() {
         return db.rawQuery("SELECT * FROM " + TestOpenHelper.TABLE_NAME, null);
     }
